requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via jsonschema
certifi==2021.10.8
    # via sentry-sdk
defusedxml==0.7.1
    # via djangorestframework-xml
django==3.2.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-parler
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==3.11.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-parler==2.3
    # via
    #   -r requirements.in
    #   django-parler-rest
django-parler-rest==2.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   django-parler-rest
    #   drf-spectacular
djangorestframework-api-key==2.1.0
    # via -r requirements.in
djangorestframework-xml==2.0.0
    # via -r requirements.in
drf-spectacular==0.21.2
    # via -r requirements.in
factory-boy==3.2.1
    # via -r requirements.in
faker==13.3.0
    # via factory-boy
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.4.0
    # via drf-spectacular
psycopg2==2.9.3
    # via -r requirements.in
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via faker
pytz==2021.3
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via drf-spectacular
sentry-sdk==1.5.6
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.8
    # via sentry-sdk
whitenoise==6.0.0
    # via -r requirements.in
